Refactor SDK CLI commands into dedicated modules
