+++ sdk/longlink/cli/__init__.py
@@ -1,2 +1,4 @@
-from .init import setup
 from .build import build_app, build_command
+from .dev import dev_command, load_app
+from .init import init_command, setup
+from .main import main
